Fixed flask tests and unified tests for all db in one file
